Add a janky little author lookup
The idea here is if you get multiple compelling results back for an
author, you can potentially pick one by also searching for a title. If
one of the authors happens to have a matching title, you can use that to
be certain.  This thing is slow AF though, and I'm not sure if I'm
using the LOC data correctly / well